Set up docker user with a password and read-only access
Also, allow only the docker user to connect from outside localhost.
This is set by a configuration file that is created by the entrypoint
script.  Thus it's not around for us to modify in the Dockerfile, but
it's already created, with the problematic lines in place, by the time
our install.sh runs.  So we alter the entrypoint script, already in
the docker image, with a sed script.  Dev-ops!